requirements.txt
requests==2.31.0
flask
pymongo
python-dotenv ==1.0.0